Moving new test files into dedicated directory
